appsody operator install and operator_utils.go functions (#249)
* initial code drop

* fixed issue with no operators present for watch ns check

* deploy and op install changes

* improve error messages, algorithm for get operator in namespace

* lint fixes

* copyright, logging changes

+++ cmd/deploy.go
@@ -30,7 +30,7 @@ import (
 	"github.com/spf13/cobra"
 )
 
-var configFile, namespace, operatorspace, watchspace, tag string
+var configFile, namespace, watchspace, tag string
 var generate, force, push bool
 
 var deployCmd = &cobra.Command{
@@ -43,15 +43,24 @@ generates a deployment manifest (yaml) file if one is not present, and uses it t
 			return generateDeploymentConfig()
 		}
 		// Check for the Appsody Operator
-		kargs := []string{"service/appsody-operator"}
-		_, err := KubeGet(kargs)
+
+		operatorExists, existingNamespace, operatorExistsErr := operatorExistsWithWatchspace(namespace)
+		if operatorExistsErr != nil {
+			return operatorExistsErr
+		}
+
+		//kargs := []string{"service/appsody-operator"}
+		//_, err := KubeGet(kargs)
 		// Performing the kubectl apply
-		if err != nil {
-			Warning.log("Failed to find Appsody operator. Attempting to install...")
-			err = installCmd.RunE(cmd, args)
+		if !operatorExists {
+			Debug.logf("Failed to find Appsody operator that watches namespace %s. Attempting to install...", namespace)
+			err := installCmd.RunE(cmd, args)
 			if err != nil {
 				return errors.Errorf("Failed to install Appsody operator. Exiting... %s", configFile)
 			}
+		} else {
+			Debug.logf("Operator exists in %s, watching %s ", existingNamespace, namespace)
+
 		}
 
 		exists, err := exists(configFile)
